requirements.txt
python-telegram-bot==22.4
Pillow
python-dotenv
